[CODE GEN] implementing return

+++ src/CodeGen/compilation/math.rs
@@ -79,7 +79,7 @@ macro_rules! MathVarType {
                     VarLocation::Reg(reg) => instr.add_operand(MachineOperand::Reg(reg)),
                 }
         
-                instr.add_operand(MachineOperand::Imm(node.inner2.val()));
+                instr.add_operand(MachineOperand::Imm(node.inner2.val() as i64));
 
                 match out {
                     VarLocation::Reg(reg) => instr.set_out(MachineOperand::Reg(reg)),
@@ -116,9 +116,9 @@ macro_rules! MathTypeType {
         
                 let mut instr = MachineInstr::new($mnemonic);
         
-                instr.add_operand(MachineOperand::Imm(node.inner1.val()));
+                instr.add_operand(MachineOperand::Imm(node.inner1.val() as i64));
         
-                instr.add_operand(MachineOperand::Imm(node.inner2.val()));
+                instr.add_operand(MachineOperand::Imm(node.inner2.val() as i64));
 
                 match out {
                     VarLocation::Reg(reg) => instr.set_out(MachineOperand::Reg(reg)),
